Add several assertions in the various files.
I made several assertions in order to help check if there are any bugs or errors in the system. I made assertions near the user inputs and areas where the inputs are changed to determine whether the change is correct or if the change is being implemented or not.

With the use of assertions, I hope to be able to better discover errors and bugs within the code and make it easier to rectify it.