Include assert to document important assumptions
Include Java's assert feature to document important
assumptions that should hold at various points in the code.

The goal is to use assertions to check and document
important assumptions about certain arguments.

Additionally, in the process of searching for locations to
include assertions. A location where assertion could be
used was found in the DeleteTaskCommand and
UpdateTaskMarkCommand classes but instead of using
assertions, using exception handling is determined to be
more suitable.