feat: Add assertions in codebase
Currently, there are no assertions within the codebase, which is not ideal as there are points where we do make assumptions about how the code behaves.

To correct this, Assertions are added in places deemed fit to make such assertions. For example, in the `DeleteCommand`, we assert that the index is within bounds before executing a delete. Another example would be list where the StringBuilder that we're returning does not return an empty output (even when there's no tasks).

These are fair assertions to make.