refactor: remove `CommandType` Enum
The `CommandType` Enum was being used to label the nature of commands executed.

However, these enums were not being used anywhere in the code other than to label the command. It remains a persistent issue as after going through many iterations, this enum marking was not being used anywhere in the codebase, which signifies the presence of dead code.

Removing the Enum completely allows us to reduce the codebase size whilst maintaining good meaning within the codebase for other developers.

The whole enum is wiped as it is not necessary to remain in the codebase.