Improving on the coding quality
The Parser class has a method that is too long and has too deep
nesting, making it difficult to read or modify code in the future.

Implemented the same logic with a switch statement instead.
Created separate methods for each case.

This is so that the code is structured in such a way that is easier
to read, organize and modify in the future.