src/main/java/duke/task/Task.java: add assert
The field isDone is only possible to be 'Y' which is yes or 'N'
which is no.

Java assert is used to ensure that it can only be 'Y' or 'N'.

Java assert is used instead of throwing exception because the
previous line assigns 'Y' or 'N' to isDone, and hence if isDone
is not 'Y' or 'N' at this point, that must be some mistakes
made by programmer, which is suitable to be handled with
assertion.