The code in the storage class is too deeply nested and convoluted.
It has to be changed in case that more types of tasks are added,
then we can add them easier.

The three different tasks that can be created has been abstracted
out into a method.

It is easier to segregate and organize the code and where to focus
on should we decide to look into or modify the code in the future.