Implement JavaFXML for MainWindow and DialogueBox controllers
As instructed in the increments for week 4, also to better understand
how fxml works.

Let's add MainWindow and DialogueBox fxml files. Also add resizing
functionality to the labels in dialogue box.

This allows us to directly modify the GUI using screenbuilder, and for
long text to be fully captured in labels