GUI update
Upon Starting the application, users are now greeted with a welcome message.

But there are more things that can be done to further improve the GUI
Let's
*Set a background image
*Make the imageView appear circular

Moreover there is a bug that needs to be fixed. Storage.save() should not write the taskList empty message into data/sirDuke.txt