Fix mark and unmark task
Fixed a bug where the new status of the marked and unmarked tasks was not saved to the database.

This caused the original status to be lost and replaced with a default "false".

Fixed by changing the representation of saved data in the database.

Also changed the syntax for loading and unloading data.

Added new function Database.changeLine() to allow for the changing of status within the database.