Implement undo command as extension
Implement two new classes to give the ability to undo the
most recent command.

The goal is to keep track of the commands done by the
user and allow undoing of successful commands. All
commands need to implement the undo method however
only certain commands are required to actually have some
functionality.

Also, some refactoring was done on the Task class and its
child classes to allow for easier undoing and the file system
is properly integrated with the current GUI (It was not
saving and loading after the GUI changes).