Improve the visual of dialogboxes
Add background color to the dialogue box to appear similar as a chat
bubble. User's dialogue box is in light blue color while the chatbot's
dialogue box is in light grey color.

Attempt to crop the profile image to a circle and try to add a circular
border to it but still getting a weird result now.