Update so that user knows directory or file created
Before JavaFX was implemented, System.out.print() was used to tell
users when a directory or file was created but this has become
redundant now.

Two class attributes is added to Ui.java so that they can be updated
in Storage.java load() method so that the showGreet() method in
Ui.java can be modified to tell the user if any directory or file was
created. The two attributes will be made false again in the showGreet()
method so that it does not repeat the same message to the user even if
the relevant directory or file has been created previously and exists.