Add JUnit testing and fix small errors
There are not enough Junit test files to
comprehensively test all functions that I have
created.

This may make the code unreliable as future
changes may break the codebase that would not be
easy to spot.

Add more JUnit testing, and some minor bug fixes
for the application that I found while JUnit test.

This will ensure that the code remains reliable
and it is easier for me to check my code before
committing.

No further references.