A-BetterGui Modified fxml files to improve GUI of Bobby.
Added background fill colours for user and Bobby's dialog box.

Added anchors to ensure that the window can be resized with the content
being resized nicely as well.

Clipped the image view with Circle shape so that the the user and
bobby's image looks more aesthetically pleasing.

Also modified the code logic to ensure that the data file to store the
task information is properly created if it does not exist when first
initialised.