Fix logic for GUI
The old logic is used for CLI and that logic design breaks in GUI.

To make the app work as intended in GUI, most of the main logic has to
be changed.

The Command.execute() method now takes in the Storage parameter. This
allows the app to store data when execute() is called instead of
run() in the CLI.

Since the GUI uses a different start() method from the CLI, this change
in design is to make the logic consistent for both interfaces.