Refactor Add Command execute method
The execute method was more than 30 LOC because it had a common
nested if-condition

Let's extract the nested if-condition to shorten the code. Also shift
error handling away from execute directly to Parser parse and the
respective task classes

This helps to shorten the method considerably, and also better
delineate responsibility of error handling to the correct classes.