Add the command to improve the code quality.
The parse method in the Parse file was a very long method, so I decided to finally add the command code to help improve the code quality.

With the command code added, I also managed to implement SLAP as well. This should help if I need to implement the same commands in other parts of the code.