Remove dead code from the code files.
There were several pieces of code that were no longer used when I started using the GUI, so I decided to remove it to save space and make it more readable and cleaner.

By removing the dead code, I am helping to improve the code quality of this project. There may still be some dead code that I may have missed, so I will keep an eye out for them