Tackled more error handling.
Handled the error cases for the extension function, C-Update.

Handled the error where the user states a task number bigger than the
size of the task list. (invalid task number) Also handled the error
when the user does not state the attribute to update and indicate with
"/" character. Handled these error cases by throwing BobbyException.