Add assert statements
Added assert statements to check for important points before data is
being passed to subsequent stages.
Ensured that taskList cannot be null when passed to Interpreter.
Ensured that parser, intrpr, taskList cannot be null when passed to
Storage.
Ensured that the addition of todo, deadline and event tasks incremented
the tasklist size.
Ensured that the deletion of tasks decrement the tasklist size.
Ensured that chatbot response cannot be an empty String.