Fix undo functionality and text wrapping in GUI
The undo functionality in the Undo class was not properly restoring
deleted tasks to their original positions in the task list.
Additionally, the text wrapping feature in the GUI was not handling long
texts properly.

The improper undo functionality could lead to inconsistencies in
task management, while the lack of text wrapping affects the user
experience negatively.

Let's,
*update the undo class to insert the task back into the proper position
*modify the dialogbox.fxml file to implement proper text wrapping for
the gui