Add assertions in Parser, Shon and Storage classes
The code does not have assertion statements.

Assertion statements should be used to define assumptions about the
code and verify them during runtime. Assertion failure can help
identify bugs in the code.

Add assertion statements to Parser, Shon and Storage classes. Let's:
* add null check for Action in Parser#parse
* check for existence of storage directory and storage file after
initialisation
* check the correct letter is used to represent tasks in add methods
of Storage class
* check that storage and tasklist is initialised properly in Shon.

The above checks are done as assertions and not Exceptions as the
failure of above is caused by programmer error.