Detect Duplicates
Add the ability to recognize and deal with duplicate items.
e.g., the same task added multiple times.

This was done by adding the equals() method to the task class,
and checking in the addTask() method of the TaskList to see if the
new task is a duplicate of any of the tasks already in the list.
If duplicate is detected, it is not added to the list.

This is to ensure that multiple tasks of the same entry are not added
to the TaskList since they would be redundant.