improve security: Add assert to check for incorrect input
Bug Fix: Add assert to minimise incorrect input.

Certain classes use inputs to create Task object or child of Task object. Exceptions are used to tackle errors in input by user. However, assertions help internal logic check to ensure code functions correctly. For example, assert is used to check that deadline set is in the future and not in the past.

Let's add assert statement to ensure that code logic.

This ensure the the code gives correct output to user and there is no logic error when using chatbot.