Modify most of classes to have a hierarchy structure
For example, the original packages only have two which is Ping and
Command. Renamed them with lower case to fit the standard of commit
then separate all of them into 4 different packages which are
command, exceptions, GUI, job, and a package folder for checking
time.

Implement a better design GUI to make the window more beautiful and
tidy.

A good and comfortable window will improve the user experience. So,
doing a GUI beauty improvement is necessary.

Let's
* change the font to make chatbot and user different
* add different background colors to distinguish chatbot and user