Use Assertions and fix exception message
The assert feature can document important assumptions at various
points in the code. Make the exception message clearer so the user can
understand what is wrong in the command.

The command that users input cannot be null. The number from the input
command is a positive integer. The time for an event cannot be null.
The keyword in the find command cannot be null.

The number in the input command must be a positive integer.
The exception message is changed to make that error clearer.
Add another case where the keyword from find is empty.