Improve code quality
Improve code quality can help improve readability, making the logic
easier to follow and debug.

Use blank lines to separate different groups of logic in a method.

Change the method initialize() in the MainWindow class to follow SLAP.
Move the greetingMessage out of the method to make a single level
of abstraction.

Having a single level of abstraction in a method can improve
readability.