Add help functionality
The bot previously lacked a help functionality, which made it difficult
for users to understand how to use the various features of the bot.

In this commit, a getHelpString method was added to the TaskList class.
This method returns a string that describes all the features of the
TaskList class and how to use them. The features covered include adding
tasks, adding deadlines, displaying tasks, deleting tasks, marking tasks
as done, unmarking tasks as done, and finding tasks. This method is then
called by the input handler.

Adding this help functionality improves the user-friendliness of the bot
by providing users with a quick way to learn how to use the bot's features.