Add assertion to DialogBox and MainWindow class
We need assertion to check if resources are loaded correctly in our code.

Assertion present in code allows us to know if our code results in any
error prematurely and with a convenient message to point to where
the error is.

Let's add assert to the two new FXML bridging classes we add in
Level 10 to check for important assertions in our code!

Other sources of exception are handled by earlier code.

To write better comments, visit oracle to find out more!