Improve the code quality
Many methods that overrides a method or implements an abstract method
in the parent class/interface does not have the @Override annotation.

This is bad coding practice, and does not allow the compiler to
raise warnings about incorrect overriding.

This commit changes the situation by adding @Override at
various places.