Implemented Category C extention: C-DetectDuplicates
Detect duplicates in the stored tasks list. If there is already a task exactly
same as the one to be added, the program will not add task to list and prompt
user to write new tasks.

Implemented isDuplicate(Task) in the 3 types of task class to check if task is the same
as the one specified.

Implemented hasDuplicates(Task, TaskList) to check for duplicates in the Tasklist.