Ui class has many code quality issues
This may have an adverse effect on readability. So rectify this problems.

Some methods like listHits() and listDueSoon() have too many levels of indentation. Solution is to extract some of the if statements out as helper methods.
Happy path is unclear in some methods due to if else statements checking for assumptions, so use assert instead.