BCD Extension
Task class: add priority

As part of Week 5, need to allow chatbot to assign priority to task based on user.

Add methods to allow priority to be changed based on user input

Initialise priority as 0 when new Task is created.

This allows easy change of priority for a particular task.