Fix readme and build file
The code is working great, but is having issues
in the userguide. This issue would be either
grammar errors or unclear examples.

This may cause the user to have issues
reading the userguide, and may see unnecessary
comments in the build file.

Edit the readme to better guide users on the use
of the application. Remove unnecessary comments
in the build file too.

This will allow the user a better experience
following the guide and code, without any
unnecessary comments.

No further references.