Implement JavaFX GUI
Changed Command and all subclasses to return a String for GUI to
display in addition to printing to the console.

Rectified bug in Storage and Tasklist instantiation and updating that
caused both private tasks arraylists to reference the same object
causing double updates to be displayed when adding tasks.

Added Launcher, Main, MainWindow and DialogBox classes to manage the
GUI.Also added DialogBox and MainWindow fxml files. Added images for
user and bot.