Update build.gradle before testing whether assertions work
The mainClass of the application was initially set to my main class,
which was where the program run from before the GUI was set in.
I have updated it to be the Launcher now that it is the entry point
of this program.

As I will be using gradle to run my program and check if the assert
statements work, I had to add the `run {enableAssertions = true}` to the file.

By ensuring the build.gradle file is updated, it can then allow me to work
on the assert statements thereafter.