Add assert statements
Assert statements were added to verify the assumption that a task number
is greater than 0 when the command involves directly typing in a task
number, specifically the "mark" and "unmark" commands.

An assert false statement has also been added to the last default case of the
switch case statements where an invalid command has been used.

With these assert statements, future modifications to e.g. the Parser or in processing
the indices of the commands can be more visible to the developer as AssertionErrors
with the respective messages will be reflected in the console.