Floofy class: remove unnecessary variable in runCommand function
This method returns a String, and used a variable to store the String
that is to be returned for each case, before finally returning that
variable at the end of the switch-case statement. This can cause users
to potentially 'trip-up' due to the anomaly in the data flow, since
there is no reason to assign the String to a variable before returning.

Let's remove the variable and return the String directly in each case.

This shortens the code, and improves the readability of the code as
readers no longer need to trace all the way to the bottom of the
switch-case method to find where the return statement is. It also
removes the need for break statements.