Parser class: replace default invalid block in parse method
The parse method parses the user input to determine the command. The
else case in the method assigns the parsed input to "invalid",
which is a slight data flow anomaly. This ultimately affects the
organisation of the runCommand() function in Floofy class, which
had to contain a case for a command named "invalid". This potentially
confuses readers.

Lets replace the assigning of the parsed input to "invalid" with
throwing FloofyException, which was created to be thrown when invalid
inputs are given by the user.

This allows us to make use of the behaviour of an exception, and also
removes the need to return an empty String in the runCommand() method in
Floofy, since without an "invalid" parsed input, the default case can be
used for invalid commands to simply throw the exception.