Change getInfo 2nd parameter to use enum
2nd parameter takes an int i and method returns the ith word from input.
Difficult to use as one has to remember which position the required info
is in.

Let's create an enum class for each type of info the user can give and
match the type to the position so that method can take in the enum class
as parameter and return the relevant info.