Parser class: Improve code quality
This is to prepare the code for C-Friendlier Syntax. The parser uses
absolute indices to access several components of the user's input,
which would not work if we were to allow for aliases.

Let's remove the use of magic numbers, with integers only used to access
relative index.

Let's also improve the readability of the each command's parser method
by abstracting out details of parsing into smaller helper methods.

These allow the code to be more easily extendable in future when a
greater variety of commands are accepted or when new commands
are created.