Add the extension for C-DetectDuplicates for BCD-Extension.
In order to reduce any duplicates within the program, I add the function for the TaskList.java to detect if the new task being added is already inside the list.

I override the equals method in the Deadline, Event and Todo files in order for the ArrayList to properly check if they are the same or not.

For the todo task, I had the arrayList compare only the description.

For the deadline task, I had the arrayList compare the description and the by date.

For the event task, I had the arrayList compare the description, the from date and the to date.

This should prevent any duplicates from being entered into the ArrayList, improving functionality.