Add assertions in DeleteCommand, MarkCommand and UnmarkCommand
Delete, Mark and Unmark commands have some common field validation checking behaviours.

The field validation can be captured using assertions instead of checked exceptions in production code.

Replacing the validation checks with assertions allows us to check for invalid inputs by the programmer. For example, assertions can be used to check for inputs that are out of bounds of the possible valid integers.

Let's replace all these validation checks with assertions to make the valid cases more explicit.

Using assertions is preferable over if statements and checked exceptions because it makes the validation checks more explicit to the programmer.