Made changes to GUI
There was an issue with the GUI where the output box would not load
more than 5 lines worth of tasks when prompted to list out tasks.

Fixing this would allow for a better user experience.

As a step towards improving the GUI, let's make sure the code
in the MainWindow.fxml file is properly calibrated. Doing so will
make the subsequent implementation of GUIs functional.