Fix Linux compatability with app
The application has issues being ran on the
Linux operating system. This incluides Ubuntu,
Kali Linux, etc.

This affects users as linux users are unable to
run the BobBot Application, and that makes the
application have a smaller target audience.

Fix the build gradle file to include the linux
configurtations to utilize the respective
dependency.

This will allow linux users to use the app,
opening up the range of audience.

No further references.