Add C-Help extension
Currently, users always get the instruction message if they type erroneous inputs,
regardless if they know how to rectify them.

As the instruction message is long, it runs the risk of clogging the conversation space.

So a message directing the user to type "help" for full instructions is made,
so the user can get advice only when he/she needs it.