Modify find command
Find command returns an ArrayList of matched tasks. Users can find it
confusing to mark or delete tasks from the provided response as their
indexes are not shown.

Modifying the find command to use an arraylist of <Task, Integer> pairs
allows for the tasks' indexes to be captured.

Let's use a new method in Ui to print these tasks with their original
index to streamlien user experience.

Using a separate method is preferable to having the displayList method
handle both standard list and find lists.