Add help command
There is no info on how to use bot when user first start it.

New users would not know the specific command structure and has to
resort to trial and error which is not user-friendly.

Lets add a helpCommand that executes when users enter "help"
which displays the command structure for every command available
so that users can immediately start using it.