Add Help extension
There is no help page.

Having a help page provides users with a consolidated view of all available commands and command formats. It is especially beneficial for first-time users.

Add a Help Page that is accessible with the command "help".

Users can easily access the help page with a simple command.