Refactor code
The code throughout has been refactored to improve readability and maintainability. The changes include renaming the handlers, simplifying the code, and updating the main I/O handler with fewer lines of code.

Previously, there was a lot of arrowhead in multiple lines which made it harder to understand. As such, those ocdes were refactores with more private functions wherever it was possible to modularise it.

Additionally, assertions have been added throughout the codebase. This helps to catch potential errors early and makes the code more robust.

Overall, the code is much more modular and easier to work with.

These changes contribute to making the code cleaner and easier to understand, which will facilitate future development and maintenance.